requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.0.1
python-dotenv==1.0.0
requests==2.31.0
hubspot-api-client==7.0.0
Werkzeug==2.3.7
phonenumbers==8.13.25
psycopg2-binary==2.9.9
alembic==1.12.1
Flask-Limiter==3.5.0
Flask-Bcrypt==1.0.1
gunicorn==21.2.0
